fixed database trouble, i think

+++ app/src/main/java/com/example/mappe2_s344104_s344045/Friend.java
@@ -25,7 +25,7 @@ public class Friend {
         this.phone = phone;
     }
 
-    public Friend(int _ID, String firstname, String lastname, String phone) {
+    public Friend(long _ID, String firstname, String lastname, String phone) {
         this._ID = _ID;
         this.firstname = firstname;
         this.lastname = lastname;
